update: 루트 컨텍스트, 서블릿 컨텍스트 서비스와 웹계층 분리. 트랜잭션 설정, 컴포넌트 스캔 베이스 재설정

+++ src/main/java/com/fundly/chat/model/SelBuyMsgDetailsDao.java
@@ -10,7 +10,7 @@ import java.util.List;
 @Mapper
 public interface SelBuyMsgDetailsDao {
     int insertMsg(SelBuyMsgDetailsDto selBuyMsgDetailsDto) throws Exception;
-    List<SelBuyMsgDetailsDto> loadAllMessages(int roomNum) throws Exception;
+    List<SelBuyMsgDetailsDto> loadAllMessages(Long roomNum) throws Exception;
     int deleteAll();
     int count();
 }
